pyproject.toml
[project]
name = "kharis-screening-agent"
version = "0.1.0"
description = "AI-powered company screening agent for investment analysts"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "anthropic>=0.40",
    "httpx>=0.27",
    "trafilatura>=1.12",
    "slack-sdk>=3.33",
    "aiohttp>=3.9",
    "pydantic-settings>=2.5",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[tool.ruff]
line-length = 100
target-version = "py311"
